implementer update and delete functionalities

+++ MajorProject/blog/views.py
@@ -13,21 +13,16 @@ def blogging(request):
                     'bname': request.POST['bname'],
                     'bcontent': request.POST['bcontent']
                 }
-        print(data)
         form = BlogForm(data)
-
         if form.is_valid():
-            print("Form----->", form)
             try:
                 form.save()
                 return redirect("/home")
             except:
                 pass
-
     else:
         print("Not valid form")
         form = BlogForm()
-
     return render(request, 'index.html', {'form': form})
 
 
@@ -36,17 +31,34 @@ def show(request):
     return render(request, "show.html", {'blogs': blogs})
 
 
-def update(request, id):
-    blog = Blog.objects.get(id=id)
-    form = BlogForm(request.POST, instance=blog)
-    if form.is_valid():
-        form.save()
-        # return redirect("/show")
-    # return render(request, 'edit.html', {'blog': blog})
+def update(request):
+    bid = request.POST['bid']
+    blog = Blog.objects.get(bid=bid)
+    blog.bcontent = request.POST['bcontent']
+    blog.save()
+    return redirect("/home")
+    # blog = Blog.objects.get(id=id)
+    # form = BlogForm(request.POST, instance=blog)
+    # if form.is_valid():
+    #     form.save()
+    #     # return redirect("/show")
+    # # return render(request, 'edit.html', {'blog': blog})
+    pass
 
 
 def destroy(request):
     print(id)
-    blog = Blog.objects.get(id=id)
-    blog.delete()
-    return redirect("/home")
+    if request.method == "POST":
+        print("hello world")
+        print(request.POST['bid'])
+        blog = Blog.objects.get(bid=request.POST['bid'])
+        print(blog)
+        blog.delete()
+        return redirect("/home")
+        form = BlogForm()
+        print(form)
+
+    else:
+        print("Not valid form")
+        form = BlogForm()
+    return render(request, 'index.html', {'form': form})
